intelligent-information-retrieval-ST7071CEM/pyproject.toml
[project]
name = "ir"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "bs4>=0.0.2",
    "fastapi[standard]>=0.116.1",
    "nltk>=3.9.1",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "selenium>=4.35.0",
    "spacy>=3.8.7",
    "watchfiles>=1.1.0",
]
